telepathy-glib: fix build
avoid this issue:
| rmdir: failed to remove `/srv/home/nitin/builds/build-gcc47/tmp/work/i586-poky-linux/telepathy-glib-0.18.1-r0/image/usr/bin': No such file or directory
NOTE: package telepathy-glib-0.18.1-r0: task do_install: Failed

no PR bump as no change in the output

+++ meta/recipes-connectivity/telepathy/telepathy-glib_0.18.1.bb
@@ -18,7 +18,13 @@ FILES_${PN} += "${datadir}/telepathy \
                 ${datadir}/dbus-1"
 
 do_install_append() {
-	rmdir ${D}${bindir}
-	rmdir ${D}${libexecdir}
-	rmdir ${D}${servicedir}
+	if [ -d ${D}${bindir} ] ; then
+		rmdir ${D}${bindir}
+	fi
+	if [ -d ${D}${libexecdir} ] ; then
+		rmdir ${D}${libexecdir}
+	fi
+	if [ -d ${D}${servicedir} ] ; then
+		rmdir ${D}${servicedir}
+	fi
 }
